Working & verlifying. More cleanup. Need to persist creds
